Added try except for safety in case a certain level set does not exist
Previous code could have thrown an error if one of the level sets for meshing did not exist in the density field. This is very likely for high density level sets that are attempted to mesh early on in training, since the density field would not have learned much early on and lower densities would be more common.